+++ Source/WinCopies.Util.Desktop.Shared/IBackgroundWorker.cs
@@ -15,10 +15,10 @@
  * You should have received a copy of the GNU General Public License
  * along with the WinCopies Framework.  If not, see <https://www.gnu.org/licenses/>. */
 
+#if !WinCopies2
 using System.ComponentModel;
 using System.Threading;
 
-#if !WinCopies2
 namespace WinCopies
 {
     /// <summary>